+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+- Added two new events for the promotion case of datalayer
 ## [2.5.1] - 2020-09-15
 ### Changed
 - Replace handler of `vtex:cart` event to `vtex:cartLoaded`.